Minor refactor - discounts services (#470)
This PR encompasses the refactoring of the Discount Service, introducing significant enhancements to the discount calculation and retrieval processes. A notable change is the implementation of the GetDiscountAmountProvider through the use of the Mediator pattern, enabling the calculation of discounts via a plugin mechanism. Additionally, the Mediator pattern has been applied to facilitate the retrieval of a list of applied discounts, ensuring a more streamlined and efficient process. The Discount Service has been further modularized by separating it into IDiscountValidationService and IDiscountProviderLoader. This division not only clarifies the service's responsibilities but also enhances its maintainability and scalability by allowing for more focused implementations of its constituent parts.